fix unauthenticated user to add to cart

+++ src/lib/actions/cart.ts
@@ -2,14 +2,14 @@
 
 import { cookies } from "next/headers";
 import { Cart } from "@/lib/db/cart_collection";
-import { isLoggedIn } from "@/lib/utils/auth";
+import { isLoggedIn } from "@/lib/actions/users";
 import { ObjectId } from "mongodb";
 import { revalidateTag } from "next/cache";
 
 const NEXT_PUBLIC_URL = process.env.NEXT_PUBLIC_URL;
 
 export async function getCart(): Promise<Cart> {
-  if (!isLoggedIn()) {
+  if (!await isLoggedIn()) {
     return { items: [], _id: new ObjectId(), userId: new ObjectId() }
   }
   const res = await fetch(`${NEXT_PUBLIC_URL}/api/cart`, {
@@ -29,6 +29,8 @@ export async function getCart(): Promise<Cart> {
 }
 
 export async function addToCart(productId: string, quantity: number) {
+  if (!await isLoggedIn()) return
+
   await fetch(`${NEXT_PUBLIC_URL}/api/cart/${productId}`, {
     method: 'POST',
     body: JSON.stringify({ quantity }),
@@ -41,6 +43,8 @@ export async function addToCart(productId: string, quantity: number) {
 }
 
 export async function updateCartItemQuantity(productId: string, quantity: number) {
+  if (!await isLoggedIn()) return
+
   await fetch(`${NEXT_PUBLIC_URL}/api/cart/${productId}`, {
     method: 'PUT',
     body: JSON.stringify({ quantity }),
@@ -53,6 +57,8 @@ export async function updateCartItemQuantity(productId: string, quantity: number
 }
 
 export async function removeFromCart(productId: string) {
+  if (!await isLoggedIn()) return
+
   await fetch(`${NEXT_PUBLIC_URL}/api/cart/${productId}`, {
     method: 'DELETE',
     headers: {
